Refactor original uri changes
