+++ src/azureStorageExplorer/parseUri.ts
@@ -3,10 +3,11 @@
  *  Licensed under the MIT License. See License.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as path from "path";
 import * as vscode from 'vscode';
 
 /**
- * Example uri: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resourcegroup1/providers/Microsoft.Storage/storageAccounts/storageaccount1/Blob Containers/container1/parentdir1/subdir
+ * Example uri: azurestorage:///container1/parentdir/subdir/blob?resourceId=/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/resourcegroup1/providers/Microsoft.Storage/storageAccounts/storageaccount1/Blob Containers/container1
  */
 export interface IParsedUri {
     /**
@@ -16,6 +17,13 @@ export interface IParsedUri {
     rootPath: string;
 
     /**
+     * Type of resource
+     * e.g. Blob Containers
+     * e.g. File Shares
+     */
+    resourceType: string;
+
+    /**
      * Name of container or file share
      * e.g. container1
      */
@@ -23,42 +31,58 @@ export interface IParsedUri {
 
     /**
      * Full path within container or file share
-     * e.g. parentdir1/subdir
+     * e.g. parentdir/subdir/blob
      */
     filePath: string;
 
     /**
      * Same as `filePath`, just with a delimiter at the end in case this is a directory
-     * e.g. parentdir1/subdir/
+     * e.g. parentdir/subdir/blob/
      */
     dirPath: string;
 
     /**
      * Path of parent directory within container or file share
-     * e.g. parentdir1
+     * e.g. parentdir/subdir
      */
     parentDirPath: string;
 
     /**
      * Name of file or directory
-     * e.g. subdir
+     * e.g. blob
      */
     baseName: string;
 }
 
-export function parseUri(uri: vscode.Uri | string, fileType: string): IParsedUri {
-    let path: string = uri instanceof vscode.Uri ? uri.path : uri;
-    const matches: RegExpMatchArray | null = path.match(`^(\/subscriptions\/[^\/]+\/resourceGroups\/[^\/]+\/providers\/Microsoft\.Storage\/storageAccounts\/[^\/]+\/${fileType}\/([^\/]+))\/?((.*?)\/?([^\/]*))$`);
-    if (!matches) {
-        throw new Error(`Invalid ${fileType} uri. Cannot view or modify ${uri}.`);
+export function parseUri(uri: vscode.Uri): IParsedUri {
+    const queryMatches: RegExpMatchArray | null = uri.query.match('^resourceId=(\/subscriptions\/[^\/]+\/resourceGroups\/[^\/]+\/providers\/Microsoft\.Storage\/storageAccounts\/[^\/]+\/([^\/]+)\/([^\/]+))$');
+    const pathMatches: RegExpMatchArray | null = uri.path.match('^\/([^\/]+[^\/]+)\/?((.*?)\/?([^\/]*))$');
+    if (!pathMatches || !queryMatches) {
+        throw new Error(`Invalid uri. Cannot view or modify ${uri}.`);
     } else {
         return {
-            rootPath: matches[1],
-            rootName: matches[2],
-            filePath: matches[3],
-            dirPath: matches[3] ? `${matches[3]}/` : '',
-            parentDirPath: matches[4],
-            baseName: matches[5]
+            rootPath: queryMatches[1],
+            resourceType: queryMatches[2],
+            rootName: queryMatches[3],
+            filePath: pathMatches[2],
+            dirPath: pathMatches[2] ? `${pathMatches[2]}/` : '',
+            parentDirPath: pathMatches[3],
+            baseName: pathMatches[4]
         };
+    }
+}
+
+export function idToUri(id: string): vscode.Uri {
+    const matches: RegExpMatchArray | null = id.match('^\/subscriptions\/([^\/]+)\/resourceGroups\/([^\/]+)\/providers\/Microsoft\.Storage\/storageAccounts\/([^\/]+)\/([^\/]+)\/([^\/]+)\/?(.*?)\/*$');
+    if (!matches) {
+        throw new RangeError(`Invalid id. Cannot view or modify ${id}.`);
+    } else {
+        const subscriptionName = matches[1];
+        const resourceGroupName = matches[2];
+        const storageAccountName = matches[3];
+        const resourceType = matches[4];
+        const groupNodeName = matches[5];
+        const filePath = matches[6];
+        return vscode.Uri.parse(`azurestorage:///${path.posix.join(groupNodeName, filePath)}?resourceId=/subscriptions/${subscriptionName}/resourceGroups/${resourceGroupName}/providers/Microsoft.Storage/storageAccounts/${storageAccountName}/${resourceType}/${groupNodeName}`);
     }
 }
